+++ requirements.txt
@@ -5,4 +5,5 @@
 pbr>=1.6 # Apache-2.0
 
 argparse
+Flask>=0.8
 PyGithub>=1.3.4 #GPL
