+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
-name = abc
-summary = Utility for storing of configs
+name = abc-server
+summary = Server for utility for storing of configs
 description-file = README.rst
 license = Apache Software License
 classifiers =
@@ -19,8 +19,8 @@ setup-hooks = pbr.hooks.setup_hook
 
 [entry_points]
 console_scripts =
-    abc = abc_ctl.runner:main
+    abc-server = abc_server.runner:main
 
 [files]
 packages =
-    abc
+    abc_server
