+++ tox.ini
@@ -8,7 +8,7 @@ usedevelop = True
 install_command = pip install -U {opts} {packages}
 setenv =
     VIRTUAL_ENV={envdir}
-    DISCOVER_DIRECTORY=abc/tests/unit
+    DISCOVER_DIRECTORY=abc_server/tests/unit
 deps =
     -r{toxinidir}/requirements.txt
     -r{toxinidir}/test-requirements.txt
@@ -24,5 +24,5 @@ commands =
 [flake8]
 show-source = true
 builtins = _
-exclude=.git,.tox,*egg,tools,my_dev_server/db/base.py
+exclude=.git,.tox,*egg
 select = E
